refactor: Dictionary module
